requirements.txt
requests
beautifulsoup4
Pillow
python-dotenv
